Implement hints (#53)
* hints

* fmt
