+++ expander_compiler/src/circuit/ir/hint_normalized/witness_solver.rs
@@ -11,10 +11,11 @@ impl<C: Config> WitnessSolver<C> {
         &self,
         vars: Vec<C::CircuitField>,
         public_vars: Vec<C::CircuitField>,
+        hint_registry: &mut HintRegistry<C::CircuitField>,
     ) -> Result<(Vec<C::CircuitField>, usize), Error> {
         assert_eq!(vars.len(), self.circuit.input_size());
         assert_eq!(public_vars.len(), self.circuit.num_public_inputs);
-        let mut a = self.circuit.eval_with_public_inputs(vars, &public_vars)?;
+        let mut a = self.circuit.eval_safe(vars, &public_vars, hint_registry)?;
         let res_len = a.len();
         a.extend(public_vars);
         Ok((a, res_len))
@@ -24,8 +25,10 @@ impl<C: Config> WitnessSolver<C> {
         &self,
         vars: Vec<C::CircuitField>,
         public_vars: Vec<C::CircuitField>,
+        hint_registry: &mut HintRegistry<C::CircuitField>,
     ) -> Result<Witness<C>, Error> {
-        let (values, num_inputs_per_witness) = self.solve_witness_inner(vars, public_vars)?;
+        let (values, num_inputs_per_witness) =
+            self.solve_witness_inner(vars, public_vars, hint_registry)?;
         Ok(Witness {
             num_witnesses: 1,
             num_inputs_per_witness,
@@ -40,12 +43,13 @@ impl<C: Config> WitnessSolver<C> {
         &self,
         num_witnesses: usize,
         f: F,
+        hint_registry: &mut HintRegistry<C::CircuitField>,
     ) -> Result<Witness<C>, Error> {
         let mut values = Vec::new();
         let mut num_inputs_per_witness = 0;
         for i in 0..num_witnesses {
             let (a, b) = f(i);
-            let (a, num) = self.solve_witness_inner(a, b)?;
+            let (a, num) = self.solve_witness_inner(a, b, hint_registry)?;
             values.extend(a);
             num_inputs_per_witness = num;
         }
